pyproject.toml
[project]
name = "athena"
version = "0.1.0"
description = "Athena: A continually learning RAG-powered AI customer service agent specialized for the logistics industry."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings",
    "sqlalchemy>=2.0.0",
    "langchain>=0.1.0",
    "langchain-community",
    "langchain-openai>=0.0.5",
    "sentence-transformers>=2.2.2",
    "chromadb",
    "onnxruntime<=1.21.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "httpx>=0.27.0",
    "asyncio>=3.4.3",
    "PyJWT>=2.8.0",
    "pytest>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.12.0",
    "flake8>=7.0.0",
    "isort>=5.13.2",
    "mypy>=1.8.0",
    "pytest>=8.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
[tool.uv]
index-url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"